refactor: modularize window renderers
